make tr_info.webseedCount and tr_info.trackerCount unsigned.

+++ utils/show.c
@@ -122,7 +122,7 @@ compare_files_by_name (const void * va, const void * vb)
 static void
 showInfo (const tr_info * inf)
 {
-  int i;
+  unsigned int i;
   char buf[128];
   tr_file ** files;
   int prevTier = -1;
@@ -185,10 +185,10 @@ showInfo (const tr_info * inf)
 
   printf ("\nFILES\n\n");
   files = tr_new (tr_file*, inf->fileCount);
-  for (i=0; i< (int)inf->fileCount; ++i)
+  for (i=0; i<inf->fileCount; ++i)
     files[i] = &inf->files[i];
   qsort (files, inf->fileCount, sizeof (tr_file*), compare_files_by_name);
-  for (i=0; i< (int)inf->fileCount; ++i)
+  for (i=0; i<inf->fileCount; ++i)
     printf ("  %s (%s)\n", files[i]->name, tr_formatter_size_B (buf, files[i]->length, sizeof (buf)));
   tr_free (files);
 }
@@ -217,7 +217,7 @@ tr_curl_easy_init (struct evbuffer * writebuf)
 static void
 doScrape (const tr_info * inf)
 {
-  int i;
+  unsigned int i;
 
   for (i=0; i<inf->trackerCount; ++i)
     {
